Refactored Service Tests file structure
