+++ FrannHammer.Services.Tests/MoveSearch/BaseMoveHitboxSearchPredicateServiceTest.cs
@@ -2,7 +2,7 @@
 using FrannHammer.Services.MoveSearch;
 using NUnit.Framework;
 
-namespace FrannHammer.Services.Tests
+namespace FrannHammer.Services.Tests.MoveSearch
 {
     [TestFixture]
     public class BaseMoveHitboxSearchPredicateServiceTest
